backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
anthropic
python-dotenv
pydantic
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
twilio
cryptography
